requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.3
typing_extensions~=4.8.0
future~=1.0.0
pillow~=10.2.0
pip~=22.0.2
wheel~=0.43.0
Jinja2~=3.1.2
gast~=0.5.4
lmdb~=1.4.1
pytz~=2024.1
thop~=0.1.1-2209072238
requests~=2.31.0
colorama~=0.4.6
PyYAML~=6.0.1
yapf~=0.40.2
importlib_metadata~=7.1.0
setuptools~=59.6.0
scipy~=1.13.0
mpmath~=1.3.0
sympy~=1.12
platformdirs~=4.2.1
tabulate~=0.9.0
astunparse~=1.6.3
tqdm~=4.66.4
triton~=2.2.0
torchvision~=0.17.2+cu121
packaging~=24.0
urllib3~=2.2.1
fsspec~=2023.4.0
addict~=2.4.0
MarkupSafe~=2.1.3
six~=1.16.0
pandas~=2.2.2
python-dateutil~=2.9.0.post0
basicsr~=1.4.2
certifi~=2024.2.2
imageio~=2.34.0
tifffile~=2024.2.12
ptflops~=0.7.3
scikit-image~=0.22.0
lazy_loader~=0.4
networkx~=3.2.1
scikit-video~=1.1.11
Markdown~=3.6
idna~=3.7
termcolor~=2.4.0
torchstat~=0.0.7
torchaudio~=2.2.2+cu121
Werkzeug~=3.0.3
tensorboardX~=2.6.2.2
line_profiler~=4.1.3
zipp~=3.18.1